requirements.txt
ddddocr>=1.3.1
flask
baidu-aip
python-dotenv
numpy
opencv-python
uvicorn